Make the GitInto file be the lowest priority
So the version information is now has the following
priorities: Branch > Tag > File

This makes it easier to have a default file in master,
then just branch when a new version will be released,
without having to update the version file at all.

This commit also fixes a bug in Branch version support,
and avoids generating a build warning if no tag exists.