Make GitInfo.Vsix compatible with NuGet v3
In NuGet v3 the nuget targets end up being added *before*
the VSSDK targets themselves, so all the *DependsOn stuff
doesn't work anymore, and we need to resort to Before/AfterTargets
instead.

Took the chance to also simplify the approach a bit, by
just replacing the source vsix manifest at the first
chance, which is right after it's located by the VS targets.