Allow overriding base direcctory for GitInfo run
Passing a /p:GitInfoBaseDir allows direct execution of this targets
file as if it were run by a project importing it. This allows repositories
with multiple GitInfo.txt files and generating a different version for
each, even if you don't have a project at that level.

Also allow ignoring the git cache write by specifying SkipWriteGitCache
property as 'true'.