Fixes build error when building a project containing XAML files
The XAML targets generate a temporary project file that is used
to generate the types used in the XAML, as part of the MarkupCompilePass1
pass. If the generated code files aren't passed in to the
GenerateTemporaryTargetAssembly via the @(_GeneratedCodeFiles)
item group that is collected by MarkupCompilePass1, the build fails
if any part of the code references the generated types.

The build error will also be super cryptic since the project failing
to build will be a temporary .csproj rather than the actual containing
project, where intellisense will show that nothing is wrong
whatesoever.