When specifying a path to rev-list, pass --full-history for consistency
If the path is empty or missing, this command:

    git rev-list --count "e91067c"..HEAD

would return the same value as 

    git rev-list --count --full-history "e91067c"..HEAD

However, if a path is specified at the end, the two will be different. 

So always pass in the `--full-history` since it results in consistent 
commit counts that account for all commits history between two 
revisions.