Use MSBuildRuntimeTime to discern xbuild
MSBuildRuntimeTime is never set by xbuild, but it can also be used
to differentiate .NET Framework or .NET Core (e.g. it'll be 'Core'
on .NET Core).

This ensures GitInfo works under `dotnet build` and friends on macOS
as well as Windows.

Improves upon PR #42. Thanks to @bojanrajkovic for the suggestion.