Allow overriding the base directory to use for commit counting
Previously, we'd always use the directory of the git info file as
the base for detecting commits on top of the base version. In some
advanced scenarios (i.e. shared code in a folder above the version
file, but separate product versioning), it's useful to be able to
specify an alternative base directory for the count.