Refactor git probing to support the new Windows Store distros of WSL
For these new distros, we can't probe the Linux file system directly to
detect the presence of /usr/bin/git. Instead, let's just try to run
`git --version` in WSL bash and see what happens. This is the same
approach we already take to detect git in the Windows PATH.

To make this clean, this moves all the probing for Windows into the
SetGitExe target. It also changes the semantics slightly so that if
$(GitExe) is set externally, it won't be overridden if git is detected
in the Windows PATH.