Add proper support for multi-targeting
Previously, we were just duplicating the GitInfo.targets, but there
are targets that read files relative to that file (i.e. the
GitInfo.cache.pp and .cs.pp) and that failed in multi-targeting
scenarios.

So switch to importing the main target from the multi-targeting one
instead, like other packages and the SDKs do.