Fix missing code from missing dependency
A missing feature in NuGetizer (see https://github.com/devlooped/nugetizer/pull/388) caused our dependency on ThisAssembly.Constants to be lost now that it explicitly sets itself as PrivateAssets=all to prevent analyzer propagation to referenced projects.

This scenario (a package reference with PrivateAssets=all AND Pack=true) was not supported by NuGetizer.

We now do, so this bump is enough to fix the issue.

Fixes #281